Added building targeting

+++ Assets/Scripts/Units/UnitData.cs
@@ -11,7 +11,7 @@ public class UnitData
     public int UnitHealth;
     public int UnitDamage;
     [Description("the lower the faster")]
-    public int UnitAttackSpeed;
+    public float UnitAttackSpeed;
     public int UnitAttackRange;
     public int UnitSpeed;
     public int TargetAmount;
